+++ p5-analysis/src/bin/p5-analyze.ts
@@ -24,6 +24,6 @@ libraries.command('check-collisions').action(checkCollisions);
 libraries.command('find-minimized-alternatives').action(findMinimizedAlternatives);
 libraries.command('test-import-paths').action(checkLibraryPaths);
 libraries.command('generate').action(generateLibraryPage);
-libraries.command('list').action(listLibraries);
+libraries.command('report').action(listLibraries);
 
 program.parse(process.argv);