+++ p5-analysis/tests/Library.test.ts
@@ -10,19 +10,19 @@ test('Library.inferLibrariesFromScripts', () => {
     []
   );
   expect(Library.inferFromScripts([`${dir}/loadSound.js`]).map(l => l.name)).toEqual([
-    'p5.sound'
+    'p5.sound',
   ]);
   expect(Library.inferFromScripts([`${dir}/dat.js`]).map(l => l.name)).toEqual([
-    'dat.gui'
+    'dat.gui',
   ]);
   expect(Library.inferFromScripts([`${dir}/ml5.poseNet.js`]).map(l => l.name)).toEqual([
-    'ml5.js'
+    'ml5.js',
   ]);
   expect(Library.inferFromScripts([`${dir}/p5.Pulse.js`]).map(l => l.name)).toEqual([
-    'p5.sound'
+    'p5.sound',
   ]);
   expect(Library.inferFromScripts([`${dir}/p5.Speech.js`]).map(l => l.name)).toEqual([
-    'p5.speech'
+    'p5.speech',
   ]);
 });